packages/core/pyproject.toml
[project]
name = "white-core"
version = "0.1.0"
description = "Shared Pydantic models, enums, and types for the White project"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/white_core"]
